Update base.py, if the given parameter 'variables' is a string, convert to dict
The parameter 'variables' should be a dictionary rather than a string. The current implementation triggers errors when the value of 'variables' is a string like "{\\"id\\": \\"walmart\\"}". In this case, an example error from the GraphQL server ariadne is like below:

ERROR:ariadne:Query variables must be a null or an object.